requirements.txt
# EyeSpy v3 Backend Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-multipart==0.0.9

# Computer Vision (MediaPipe — no dlib dependency)
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.13.0

# Auth
bcrypt==4.2.0
python-jose[cryptography]==3.3.0